add constraints to db and other minor fixes

+++ hotel-microservice-project/main-service/src/main/java/org/main/service/dto/UserResponseDTO.java
@@ -16,7 +16,7 @@ public class UserResponseDTO implements Serializable {
     private int id;
     private String comment;
     private Timestamp dateTimeResponse;
-    private byte rate;
+    private double rate;
     private int userId;
     private String userLogin;
     private int room;
